Update UI: Merged the revenue headers in a single row, merged time filter header and billings header as a single file.
